flydsl/gemm: simplify to NT+NN only; move persistent variants out to grouped_gemm
User-driven cleanup of gemm_fp8_kernel.py:
- Drop CRR (TN/wgrad) kernel: _compile_dense_crr (~333 LOC), S2RLoaderTrA
  loader (~60 LOC), all _CRR_* env knobs + _pick_group_m_crr helper.
  CRR dispatch branch in gemm_fp8_tensorwise_flydsl_kernel replaced with
  NotImplementedError pointing at the grouped-gemm path.
- Move persistent variants to primus_turbo/flydsl/grouped_gemm/persistent_dense.py:
  _compile_dense_nt_persistent + _compile_dense_nn_persistent + their
  _pick_persist_factor* helpers + opt-in env knobs (FLYDSL_RCR_PERSIST_FACTOR,
  FLYDSL_NN_PERSIST_FACTOR, FLYDSL_RCR_A11_SAMETILE, FLYDSL_NN_B7_SAMETILE).
  These were default-off + BLOCKED on iter>0 wrong-output bug (post-pass07
  / LLVM AMDGPU / HW) so dead in dense path; kept as baseline for future
  grouped-gemm persistent-thread + work-stealing across experts.
- Simplify dispatch: NN native path + NT/TT canonicalised path only.
  TT remains supported via host transpose to NT (no behavior change).
- Module docstring updated to reflect NT+NN+TT-only scope.

Net file sizes:
  gemm_fp8_kernel.py    2376 → 1121 (−53%)
  persistent_dense.py   new  729  (extracted)
