+++ primus_turbo/flydsl/gemm/gemm_fp8_kernel.py
@@ -28,7 +28,7 @@ Constraints:
   - K % 128 == 0     (BLOCK_K=128, K_ITERS compile-time constant, K_ITERS >= 2)
   - out_dtype == torch.bfloat16  (StoreC fixed)
   - per-tensor scale  (a_scale / b_scale scalar fp32, broadcast inside wrapper)
-  - All four (trans_a, trans_b) combos supported; trans_c via post-hoc transpose
+  - NT, NN, TT supported (TT via host transpose to NT); trans_c via post-hoc transpose
 """
 
 import functools
@@ -390,373 +390,7 @@ if _FLYDSL_OK:
 
         return launch_dense_nt
 
-    # ──────────────────────────────────────────────────────────────────────
-    # Persistent NT variant (Session A7 prototype, A6.1 follow-on).
-    #
-    # A6 PMC diagnosis (RCR_A6_DIAGNOSTIC.md) identified kernel-launch / wave-init
-    # overhead as the dominant remaining gap vs hipblaslt: hbl runs ~constant
-    # ~215K wave/launch (persistent + work-steal), fly runs (num_tiles * 8) wave
-    # /launch (1-WG-per-tile non-persistent), 32x ratio on qwen_down_B16_M4096
-    # drives 1.30x kernel-only gap. Source/scheduler/swizzle/tile-autotune
-    # levers (A2-A5) all 0 effect; persistent restructure is the only remaining
-    # lever per A6 PMC chain.
-    #
-    # MVP design (this session):
-    #   - constexpr PERSIST_FACTOR (PF) wraps full per-tile body in
-    #     range_constexpr(PF) — IR is unrolled PF copies, no scf.for state
-    #     plumbing. Trades code-size growth (PF*) for WG-launch reduction (PF*).
-    #   - Launcher: grid_x = ceildiv(num_tiles, PF). Each WG owns PF
-    #     consecutive tiles: pid = block_idx.x * PF + tile_iv.
-    #   - Requires num_tiles % PF == 0 (host-side gate; falls back to
-    #     non-persistent variant otherwise).
-    #   - Per-iter: re-bind LDS ptrs (Python swaps in prior iter), reset acc
-    #     fragments, full prelude → main K-loop → epilog 1 → epilog 2 → store.
-    #   - Cross-iter sync: rocdl.s_barrier() before next iter's prelude to
-    #     ensure prior tile's outstanding LDS state is settled before the
-    #     same LDS slots are rewritten.
-    # ──────────────────────────────────────────────────────────────────────
 
-    @functools.lru_cache(maxsize=128)
-    def _compile_dense_nt_persistent(
-        K: int, BLOCK_M: int = 256, BLOCK_N: int = 256, PERSIST_FACTOR: int = 2,
-        A11_SAMETILE: bool = False,
-    ):
-        BLOCK_K = 128
-        assert BLOCK_M >= 128 and BLOCK_N >= 256 and BLOCK_M % 128 == 0 and BLOCK_N % 256 == 0
-        assert K % BLOCK_K == 0
-        assert PERSIST_FACTOR >= 1
-
-        K_ITERS = K // BLOCK_K
-        assert K_ITERS >= 2
-
-        N_TILES_A = BLOCK_M // 64
-        N_TILES_B = BLOCK_N // 128
-        N_ACCUMS = N_TILES_A * N_TILES_B
-
-        LDS_BLOCK_M = BLOCK_M // 2
-        LDS_BLOCK_N = BLOCK_N // 2
-
-        N_LDS_STEPS_A = LDS_BLOCK_M // 64
-        N_LDS_STEPS_B = LDS_BLOCK_N // 64
-        N_LDS_ROUNDS = max(N_LDS_STEPS_A, N_LDS_STEPS_B)
-
-        a_lds_size = LDS_BLOCK_M * BLOCK_K
-        b_lds_size = LDS_BLOCK_N * BLOCK_K
-
-        @fx.struct
-        class SharedStorage:
-            A_lds_cur_0: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            A_lds_cur_1: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            A_lds_next_0: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            A_lds_next_1: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            B_lds_cur_0: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-            B_lds_cur_1: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-            B_lds_next_0: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-            B_lds_next_1: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-
-        @flyc.kernel(known_block_size=[512, 1, 1])
-        def kernel_dense_nt_persistent(
-            A: fx.Tensor,
-            B_T: fx.Tensor,
-            C: fx.Tensor,
-            A_scale: fx.Tensor,
-            B_scale: fx.Tensor,
-            c_m: fx.Int32,
-            c_n: fx.Int32,
-        ):
-            F8_IR_t = fx.Float8E4M3FN.ir_type
-
-            n_blocks = ceildiv(c_n, BLOCK_N)
-
-            lds = fx.SharedAllocator().allocate(SharedStorage).peek()
-
-            lane_id = fx.thread_idx.x % 64
-            wave_id = fx.thread_idx.x // 64
-            wave_m = wave_id // 4
-            wave_n = wave_id % 4
-
-            gA = make_fp8_buffer_tensor(A, F8_IR_t)
-            gB = make_fp8_buffer_tensor(B_T, F8_IR_t)
-            a_div = fx.logical_divide(gA, fx.make_layout(1, 1))
-            b_div = fx.logical_divide(gB, fx.make_layout(1, 1))
-
-            gl_off_a = compute_global_swizzle(lane_id, wave_id, K, N_LDS_ROUNDS, preshuffled=False)
-            gl_off_b = compute_global_swizzle(lane_id, wave_id, K, N_LDS_ROUNDS, preshuffled=False)
-
-            # Fix #2: instantiate ALL helpers per iter — rule out cached SSA
-            # values in copy_atom / mma_atom / rmem_tensor objects shared
-            # across iters. Fix #1 (StoreC-only per-iter) was insufficient.
-
-            # A10 (A9.1 follow-up): outer persistent tile loop emitted as
-            # scf.for (runtime loop, single IR body copy) instead of
-            # Python-unroll (PF IR body copies). A7-A9 proved odd-pid tiles
-            # carry a 0.14-0.21 abs error in the Python-unroll variant; A8
-            # IR dump + A9 source-level fixes (StoreC per-iter, ALL helpers
-            # per-iter) all failed to dislodge the bug → bug is post-pass07
-            # (LLVM AMDGPU backend regalloc/scheduler across doubled body
-            # without scf.for region boundary, or pass07/08 SSA merge on
-            # doubled body). scf.for gives LLVM an explicit region; backend
-            # cannot reuse VGPRs across iters. Single IR body copy also
-            # halves kernel binary size + speeds compile.
-            #
-            # AST rewriter contract (FlyDSL ast_rewriter.py
-            # _transform_for_auto): `for tile_iv in range(N):` (NOT
-            # `tuple(range(N))` and NOT `range_constexpr(N)`) triggers
-            # scf.for lowering. Body-local variables (mfma, a_g2s, a_cur0,
-            # c00_frag, etc) are emitted as scf.for body code (NOT
-            # loop-carried iter_args) because they are not assigned before
-            # the for-loop. iv `tile_iv` is exposed as Int32 i32 SSA.
-            for tile_iv in range(PERSIST_FACTOR):
-                mfma = Mfma16x16x128(N_TILES_A, N_TILES_B)
-                a_g2s = G2SLoader(a_div, gl_off_a, N_LDS_STEPS_A, F8_IR_t, wave_id)
-                b_g2s = G2SLoader(b_div, gl_off_b, N_LDS_STEPS_B, F8_IR_t, wave_id)
-                a_s2r = S2RLoader(wave_m, N_TILES_A)
-                b_s2r = S2RLoader(wave_n, N_TILES_B)
-                store_c = StoreC(A_scale, B_scale, C, c_m, c_n, mfma.idx, N_TILES_A, N_TILES_B)
-                # Re-bind LDS slot ptrs each iter (inner main-loop Python ptr
-                # swaps don't leak across constexpr-unrolled outer iters).
-                a_cur0 = lds.A_lds_cur_0
-                a_cur1 = lds.A_lds_cur_1
-                a_next0 = lds.A_lds_next_0
-                a_next1 = lds.A_lds_next_1
-                b_cur0 = lds.B_lds_cur_0
-                b_cur1 = lds.B_lds_cur_1
-                b_next0 = lds.B_lds_next_0
-                b_next1 = lds.B_lds_next_1
-
-                # Each WG owns PF consecutive tiles (sequential-stride);
-                # PF=2 host-side gate ensures num_tiles % PF == 0.
-                # A11 sametile probe (FLYDSL_RCR_A11_SAMETILE=1): force iter1
-                # to process iter0's tile by dropping tile_iv from pid. The
-                # scf.for loop body STILL runs PF times (tile_iv is still i32
-                # SSA iv, just unused for coord derivation); iter1 overwrites
-                # iter0's same output. Discriminates two hypotheses:
-                #   (a) iter1 wrong because tile_iv → pid → coords IR alias
-                #       → sametile output ≡ PF=0 baseline (✓ bit-exact)
-                #   (b) iter1 wrong because WG-state pollution (LDS/VGPR/AGPR
-                #       /scoreboard from iter0)
-                #       → sametile output STILL has 0.21 abs error
-                # Python-level ternary: A11_SAMETILE is closure-const Python
-                # bool, evaluated at trace time. Avoids AST rewriter
-                # converting `if A11_SAMETILE:` into scf.if (which scopes
-                # branch-local vars and breaks subsequent uses of pid).
-                tile_iv_for_pid = 0 if A11_SAMETILE else tile_iv
-                pid = fx.block_idx.x * PERSIST_FACTOR + tile_iv_for_pid
-                block_m = pid // n_blocks
-                block_n = pid % n_blocks
-
-                A0_gl_offset = (block_m * BLOCK_M) * K
-                A1_gl_offset = (block_m * BLOCK_M + LDS_BLOCK_M) * K
-                B0_gl_offset = (block_n * BLOCK_N) * K
-                B1_gl_offset = (block_n * BLOCK_N + LDS_BLOCK_N) * K
-
-                c00_frag = [mfma.zero_value] * N_ACCUMS
-                c01_frag = [mfma.zero_value] * N_ACCUMS
-                c10_frag = [mfma.zero_value] * N_ACCUMS
-                c11_frag = [mfma.zero_value] * N_ACCUMS
-
-                # Cross-tile sync. A7-A9 + A10.S2 + A10.H1 (60-cycle s_nop
-                # after full drain) all verified: v/lgkm/exp full drain
-                # +/- mfma-writeback s_nop padding does NOT fix iter>0
-                # wrong-output bug. Bug is robust across Python-unroll AND
-                # scf.for runtime-loop variants — root cause is post-pass07
-                # (HW or LLVM AMDGPU backend) and not source-level drainable.
-                # Drain kept as best-effort sync (mirrors implicit kernel-exit
-                # drain of non-persistent baseline). No-op semantically on
-                # first iter (clean state).
-                _llvm.inline_asm(
-                    res=None,
-                    operands_=[],
-                    asm_string="s_waitcnt vmcnt(0) lgkmcnt(0) expcnt(0)\ns_barrier",
-                    constraints="",
-                    has_side_effects=True,
-                )
-
-                # Prelude (identical to baseline non-persistent).
-                b_g2s.load(b_cur0, B0_gl_offset + 0 * BLOCK_K)
-                a_g2s.load(a_cur0, A0_gl_offset + 0 * BLOCK_K)
-                b_g2s.load(b_cur1, B1_gl_offset + 0 * BLOCK_K)
-                a_g2s.load(a_cur1, A1_gl_offset + 0 * BLOCK_K)
-
-                if wave_m == 1:
-                    rocdl.s_barrier()
-
-                wait_barrier(N_LDS_STEPS_A + N_LDS_STEPS_B)
-
-                b_g2s.load(b_next0, B0_gl_offset + 1 * BLOCK_K)
-                a_g2s.load(a_next0, A0_gl_offset + 1 * BLOCK_K)
-                b_g2s.load(b_next1, B1_gl_offset + 1 * BLOCK_K)
-
-                wait_barrier(N_LDS_STEPS_A + 2 * N_LDS_STEPS_B)
-
-                # Ensure all prelude buffer_load_lds → LDS WRITES are committed
-                # before K-loop's ds_read. wait_barrier() only drains vmcnt;
-                # buffer_load_lds LDS-write half is tracked by lgkmcnt and is
-                # NOT drained otherwise (baseline kernel-launch wave-init delay
-                # hides this, but persistent has no such delay between tiles).
-                _llvm.inline_asm(
-                    res=None,
-                    operands_=[],
-                    asm_string="s_waitcnt lgkmcnt(0)",
-                    constraints="",
-                    has_side_effects=True,
-                )
-
-                # Main K-loop (identical to baseline).
-                for k in range_constexpr(K_ITERS - 2):
-                    b0_frag = b_s2r.load(b_cur0)
-                    a0_frag = a_s2r.load(a_cur0)
-                    a_g2s.load(a_next1, A1_gl_offset + (k + 1) * BLOCK_K)
-                    rocdl.s_barrier()
-
-                    rocdl.s_setprio(1)
-                    c00_frag = mfma.call(a0_frag, b0_frag, c00_frag)
-                    rocdl.s_setprio(0)
-                    rocdl.s_barrier()
-
-                    b1_frag = b_s2r.load(b_cur1)
-                    b_g2s.load(b_cur0, B0_gl_offset + (k + 2) * BLOCK_K)
-                    rocdl.s_barrier()
-
-                    rocdl.s_setprio(1)
-                    c01_frag = mfma.call(a0_frag, b1_frag, c01_frag)
-                    rocdl.s_setprio(0)
-                    rocdl.s_barrier()
-
-                    a1_frag = a_s2r.load(a_cur1)
-                    a_g2s.load(a_cur0, A0_gl_offset + (k + 2) * BLOCK_K)
-                    rocdl.s_barrier()
-
-                    rocdl.s_setprio(1)
-                    c10_frag = mfma.call(a1_frag, b0_frag, c10_frag)
-                    rocdl.s_setprio(0)
-                    rocdl.s_barrier()
-
-                    b_g2s.load(b_cur1, B1_gl_offset + (k + 2) * BLOCK_K)
-                    wait_barrier(2 * N_LDS_STEPS_A + N_LDS_STEPS_B)
-
-                    rocdl.s_setprio(1)
-                    c11_frag = mfma.call(a1_frag, b1_frag, c11_frag)
-                    rocdl.s_setprio(0)
-                    rocdl.s_barrier()
-
-                    a_cur0, a_next0 = a_next0, a_cur0
-                    a_cur1, a_next1 = a_next1, a_cur1
-                    b_cur0, b_next0 = b_next0, b_cur0
-                    b_cur1, b_next1 = b_next1, b_cur1
-
-                # Epilog 1 (identical to baseline, stale-a1 fix included).
-                k = K_ITERS - 2
-                b0_frag = b_s2r.load(b_cur0)
-                a0_frag = a_s2r.load(a_cur0)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c00_frag = mfma.call(a0_frag, b0_frag, c00_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                b1_frag = b_s2r.load(b_cur1)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c01_frag = mfma.call(a0_frag, b1_frag, c01_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                a1_frag = a_s2r.load(a_cur1)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c10_frag = mfma.call(a1_frag, b0_frag, c10_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                b0_frag = b_s2r.load(b_next0)
-                a_g2s.load(a_next1, A1_gl_offset + (k + 1) * BLOCK_K)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c11_frag = mfma.call(a1_frag, b1_frag, c11_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                a_cur0, a_next0 = a_next0, a_cur0
-                a_cur1, a_next1 = a_next1, a_cur1
-                b_cur0, b_next0 = b_next0, b_cur0
-                b_cur1, b_next1 = b_next1, b_cur1
-
-                # Epilog 2.
-                a0_frag = a_s2r.load(a_cur0)
-                wait_barrier(0)
-
-                rocdl.s_setprio(1)
-                c00_frag = mfma.call(a0_frag, b0_frag, c00_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                b1_frag = b_s2r.load(b_cur1)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c01_frag = mfma.call(a0_frag, b1_frag, c01_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                a1_frag = a_s2r.load(a_cur1)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c10_frag = mfma.call(a1_frag, b0_frag, c10_frag)
-                c11_frag = mfma.call(a1_frag, b1_frag, c11_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                wave_n_offset = wave_n * (N_TILES_B * 16)
-                wave_m_offset = wave_m * (N_TILES_A * 16)
-                base_row = block_m * BLOCK_M + wave_m_offset
-                base_col = block_n * BLOCK_N + wave_n_offset
-
-                store_c.store(c00_frag, base_row + 0, base_col + 0)
-                store_c.store(c01_frag, base_row + 0, base_col + LDS_BLOCK_N)
-                store_c.store(c10_frag, base_row + LDS_BLOCK_M, base_col + 0)
-                store_c.store(c11_frag, base_row + LDS_BLOCK_M, base_col + LDS_BLOCK_N)
-
-        @flyc.jit
-        def launch_dense_nt_persistent(
-            A: fx.Tensor,
-            B_T: fx.Tensor,
-            C: fx.Tensor,
-            A_scale: fx.Tensor,
-            B_scale: fx.Tensor,
-            c_m: fx.Int32,
-            c_n: fx.Int32,
-            stream: fx.Stream,
-        ):
-            num_tiles = ceildiv(c_m, BLOCK_M) * ceildiv(c_n, BLOCK_N)
-            grid_x = ceildiv(num_tiles, PERSIST_FACTOR)
-            kernel_dense_nt_persistent(
-                A,
-                B_T,
-                C,
-                A_scale,
-                B_scale,
-                c_m,
-                c_n,
-                value_attrs={"rocdl.waves_per_eu": 2, "rocdl.flat_work_group_size": "512,512"},
-            ).launch(grid=(grid_x, 1, 1), block=(512, 1, 1), stream=stream)
-
-        return launch_dense_nt_persistent
-
-    # ──────────────────────────────────────────────────────────────────────
-    # NN-layout dense fp8 kernel (used by dgrad path).
-    #
-    # A is [M, K] row-major K-contig (same as NT). B is [K, N] row-major --
-    # contraction sweeps DOWN B's rows so loading B as K-contig per N-col is
-    # strided in gmem (we still use BufferCopyLDS128b N-contig load -> N-contig
-    # LDS) and the LDS->register path uses ds_read_b64_tr_b8 which HW-transposes
-    # 8 K-rows × 8 N-cols per call so each lane ends up with K-contig fragments
-    # in the mfma B-operand format.
     # ──────────────────────────────────────────────────────────────────────
 
     _LDS_PTR_TYPE = None
@@ -879,66 +513,6 @@ if _FLYDSL_OK:
                 #   v8i32[2..3] = call 1 → bytes 8..15
                 #   v8i32[4..5] = call 2 → bytes 16..23 (K + 64 jump)
                 #   v8i32[6..7] = call 3 → bytes 24..31
-                v4_lo = calls[0].shuffle(calls[1], [0, 1, 2, 3])
-                v4_hi = calls[2].shuffle(calls[3], [0, 1, 2, 3])
-                frag.append(v4_lo.shuffle(v4_hi, list(range(8))))
-            return frag
-
-    class S2RLoaderTrA:
-        """LDS -> mfma A-operand transpose-load. Mirrors `S2RLoaderTr` but the
-        per-wave LDS column stride is parameterized as `n_tiles_a * 16` (not
-        the hardcoded 32 of `S2RLoaderTr`).  The hardcoded `wave_n * 32`
-        happened to equal `wave_n * (n_tiles_b * 16)` for the NN B case where
-        `n_tiles_b == 2`; for CRR A with `n_tiles_a == 4` (BLOCK_M=256,
-        wave_m ∈ [0,2)) the correct stride is 64 so the 2 wave_m's tile the
-        LDS_BLOCK_M=128 K_out cols without overlap or gaps.
-
-        Lane / byte convention identical to S2RLoaderTr (mfma A and B operands
-        share per-lane data format for mfma_f32_16x16x128_f8f6f4).
-        """
-
-        _K_BASE = (0, 8, 64, 72)
-        TR_TYPE = None
-
-        def __init__(self, wave_idx, n_tiles_a):
-            self.wave_idx = wave_idx
-            self.n_tiles_a = n_tiles_a
-            self.lane_id = fx.thread_idx.x % 64
-
-        def load(self, lds_src, preshuffled=False):
-            assert not preshuffled, "S2RLoaderTrA does not support preshuffled"
-            if S2RLoaderTrA.TR_TYPE is None:
-                S2RLoaderTrA.TR_TYPE = Vec.make_type(2, fx.Int32)
-            tr_type = S2RLoaderTrA.TR_TYPE
-
-            base_i32 = fx.Int32(fx.ptrtoint(lds_src.ptr))
-            I = self.lane_id // 16
-            L_in_sg = self.lane_id % 16
-            wave_col_stride = self.n_tiles_a * 16
-
-            frag = []
-            for tile_i in range_constexpr(self.n_tiles_a):
-                calls = []
-                for c in range_constexpr(4):
-                    K_log = I * 16 + S2RLoaderTrA._K_BASE[c] + (L_in_sg // 2)
-                    r_step = K_log // 64
-                    W = (K_log % 64) // 8
-                    K_mod_8 = K_log % 8
-                    swz_K = ((K_log % 16) // 2) * 16
-                    tile_col_start = self.wave_idx * wave_col_stride + tile_i * 16
-                    j_chunk = (tile_col_start // 16) ^ (swz_K // 16)
-                    ptr_offset = (
-                        W * 1024
-                        + r_step * 8192
-                        + K_mod_8 * 128
-                        + j_chunk * 16
-                        + (L_in_sg % 2) * 8
-                    )
-                    ptr_i32 = base_i32 + fx.Int32(ptr_offset)
-                    ptr = _lds_ptr_from_i32(ptr_i32)
-                    v = rocdl.ds_read_tr8_b64(tr_type, ptr).result
-                    calls.append(Vec(v))
-
                 v4_lo = calls[0].shuffle(calls[1], [0, 1, 2, 3])
                 v4_hi = calls[2].shuffle(calls[3], [0, 1, 2, 3])
                 frag.append(v4_lo.shuffle(v4_hi, list(range(8))))
@@ -1274,690 +848,9 @@ if _FLYDSL_OK:
 
         return launch_dense_nn
 
-    # ──────────────────────────────────────────────────────────────────────
-    # Session B7: persistent NN variant — sametile probe to confirm/refute
-    # whether the A11 iter>0 race (WG-state pollution) reproduces in the
-    # NN layout. Structural mirror of `_compile_dense_nt_persistent`:
-    #   - scf.for outer loop (`for tile_iv in range(PERSIST_FACTOR)`)
-    #   - all helpers (mfma, g2s, s2r, store_c) instantiated per iter
-    #   - LDS slot ptrs re-bound each iter
-    #   - cross-iter drain ASM (vmcnt/lgkmcnt/expcnt + s_barrier)
-    #   - prelude → main K-loop → epilog 1 → epilog 2 → store
-    #
-    # NN-specific differences from NT persistent:
-    #   - B offset is column-based: `block_n * BLOCK_N + {0, LDS_BLOCK_N}`
-    #     (no `*K` because B is [K, N] row-major)
-    #   - B K-iter step uses `(k + x) * BLOCK_K * c_n` (stride-by-c_n rows)
-    #   - B uses `compute_global_swizzle_nn(lane_id, wave_id, c_n, ...)`
-    #   - B s2r uses `S2RLoaderTr` (ds_read_b64_tr_b8 transpose-load)
-    #   - `_ = str(fx.thread_idx.x)` workaround retained (flydsl trace-order
-    #     bug for transposed B reads when thread_idx not pre-materialized)
-    #
-    # B7 sametile probe: when B7_SAMETILE=True AND PF>=2, iter1+ drops
-    # tile_iv from pid (forces same coord as iter0). Same discriminator as
-    # A11:
-    #   - bit-exact vs PF=0 baseline → tile_iv→pid IR alias (hypothesis a)
-    #   - residual abs error      → WG-state pollution (hypothesis b)
-    # NN replication of A11 verdict expected with high probability (same
-    # HW, same 8-wave WG, same mfma scoreboard, similar LDS slot reuse
-    # pattern); but explicit probe required for B7 closure.
-    # ──────────────────────────────────────────────────────────────────────
-
-    @functools.lru_cache(maxsize=128)
-    def _compile_dense_nn_persistent(
-        K: int, BLOCK_M: int = 256, BLOCK_N: int = 256, PERSIST_FACTOR: int = 2,
-        B7_SAMETILE: bool = False,
-    ):
-        BLOCK_K = 128
-        assert BLOCK_M >= 128 and BLOCK_N >= 256 and BLOCK_M % 128 == 0 and BLOCK_N % 256 == 0
-        assert K % BLOCK_K == 0
-        assert PERSIST_FACTOR >= 1
-
-        K_ITERS = K // BLOCK_K
-        assert K_ITERS >= 2
-
-        N_TILES_A = BLOCK_M // 64
-        N_TILES_B = BLOCK_N // 128
-        N_ACCUMS = N_TILES_A * N_TILES_B
-
-        LDS_BLOCK_M = BLOCK_M // 2
-        LDS_BLOCK_N = BLOCK_N // 2
-
-        N_LDS_STEPS_A = LDS_BLOCK_M // 64
-        N_LDS_STEPS_B = LDS_BLOCK_N // 64
-        N_LDS_ROUNDS = max(N_LDS_STEPS_A, N_LDS_STEPS_B)
-
-        a_lds_size = LDS_BLOCK_M * BLOCK_K
-        b_lds_size = LDS_BLOCK_N * BLOCK_K
-
-        @fx.struct
-        class SharedStorage:
-            A_lds_cur_0: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            A_lds_cur_1: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            A_lds_next_0: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            A_lds_next_1: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            B_lds_cur_0: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-            B_lds_cur_1: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-            B_lds_next_0: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-            B_lds_next_1: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-
-        @flyc.kernel(known_block_size=[512, 1, 1])
-        def kernel_dense_nn_persistent(
-            A: fx.Tensor,
-            B: fx.Tensor,
-            C: fx.Tensor,
-            A_scale: fx.Tensor,
-            B_scale: fx.Tensor,
-            c_m: fx.Int32,
-            c_n: fx.Int32,
-        ):
-            # See `_compile_dense_nn` docstring: forces thread_idx.x
-            # materialization before S2RLoaderTr lazy-eval inside loops.
-            _ = str(fx.thread_idx.x)
-            F8_IR_t = fx.Float8E4M3FN.ir_type
-
-            n_blocks = ceildiv(c_n, BLOCK_N)
-
-            lds = fx.SharedAllocator().allocate(SharedStorage).peek()
-
-            lane_id = fx.thread_idx.x % 64
-            wave_id = fx.thread_idx.x // 64
-            wave_m = wave_id // 4
-            wave_n = wave_id % 4
-
-            gA = make_fp8_buffer_tensor(A, F8_IR_t)
-            gB = make_fp8_buffer_tensor(B, F8_IR_t)
-            a_div = fx.logical_divide(gA, fx.make_layout(1, 1))
-            b_div = fx.logical_divide(gB, fx.make_layout(1, 1))
-
-            gl_off_a = compute_global_swizzle(lane_id, wave_id, K, N_LDS_ROUNDS, preshuffled=False)
-            gl_off_b = compute_global_swizzle_nn(lane_id, wave_id, c_n, N_LDS_ROUNDS)
-
-            # Mirror NT A10/A11 design: scf.for outer loop, all helpers per
-            # iter to rule out cached SSA values in copy_atom / mma_atom
-            # objects across iters.
-            for tile_iv in range(PERSIST_FACTOR):
-                mfma = Mfma16x16x128(N_TILES_A, N_TILES_B)
-                a_g2s = G2SLoader(a_div, gl_off_a, N_LDS_STEPS_A, F8_IR_t, wave_id)
-                b_g2s = G2SLoader(b_div, gl_off_b, N_LDS_STEPS_B, F8_IR_t, wave_id)
-                a_s2r = S2RLoader(wave_m, N_TILES_A)
-                b_s2r = S2RLoaderTr(wave_n, N_TILES_B)
-                store_c = StoreC(A_scale, B_scale, C, c_m, c_n, mfma.idx, N_TILES_A, N_TILES_B)
-
-                a_cur0 = lds.A_lds_cur_0
-                a_cur1 = lds.A_lds_cur_1
-                a_next0 = lds.A_lds_next_0
-                a_next1 = lds.A_lds_next_1
-                b_cur0 = lds.B_lds_cur_0
-                b_cur1 = lds.B_lds_cur_1
-                b_next0 = lds.B_lds_next_0
-                b_next1 = lds.B_lds_next_1
-
-                # B7 sametile probe (mirror of A11): drop tile_iv from pid
-                # so iter1+ overwrites iter0's tile with same coord. The
-                # scf.for body STILL runs PF times. See A11 verdict
-                # discriminator in `_compile_dense_nt_persistent`.
-                tile_iv_for_pid = 0 if B7_SAMETILE else tile_iv
-                pid = fx.block_idx.x * PERSIST_FACTOR + tile_iv_for_pid
-                block_m = pid // n_blocks
-                block_n = pid % n_blocks
-
-                # A: same as NT (row-major M-major stride).
-                A0_gl_offset = (block_m * BLOCK_M) * K
-                A1_gl_offset = (block_m * BLOCK_M + LDS_BLOCK_M) * K
-                # B: NN-specific column-base offsets.
-                B0_gl_offset = block_n * BLOCK_N + 0
-                B1_gl_offset = block_n * BLOCK_N + LDS_BLOCK_N
-
-                c00_frag = [mfma.zero_value] * N_ACCUMS
-                c01_frag = [mfma.zero_value] * N_ACCUMS
-                c10_frag = [mfma.zero_value] * N_ACCUMS
-                c11_frag = [mfma.zero_value] * N_ACCUMS
-
-                # Cross-tile sync (best-effort drain; A7-A12 proved source
-                # drain does NOT fix the iter>0 race in NT — included here
-                # to match NT persistent skeleton 1:1).
-                _llvm.inline_asm(
-                    res=None,
-                    operands_=[],
-                    asm_string="s_waitcnt vmcnt(0) lgkmcnt(0) expcnt(0)\ns_barrier",
-                    constraints="",
-                    has_side_effects=True,
-                )
-
-                # Prelude (identical to NN baseline non-persistent).
-                b_g2s.load(b_cur0, B0_gl_offset + 0 * BLOCK_K * c_n)
-                a_g2s.load(a_cur0, A0_gl_offset + 0 * BLOCK_K)
-                b_g2s.load(b_cur1, B1_gl_offset + 0 * BLOCK_K * c_n)
-                a_g2s.load(a_cur1, A1_gl_offset + 0 * BLOCK_K)
-
-                if wave_m == 1:
-                    rocdl.s_barrier()
-
-                wait_barrier(N_LDS_STEPS_A + N_LDS_STEPS_B)
-
-                b_g2s.load(b_next0, B0_gl_offset + 1 * BLOCK_K * c_n)
-                a_g2s.load(a_next0, A0_gl_offset + 1 * BLOCK_K)
-                b_g2s.load(b_next1, B1_gl_offset + 1 * BLOCK_K * c_n)
-
-                wait_barrier(N_LDS_STEPS_A + 2 * N_LDS_STEPS_B)
-
-                # Ensure prelude buffer_load_lds → LDS WRITES committed
-                # before K-loop's ds_read (mirror NT persistent fix).
-                _llvm.inline_asm(
-                    res=None,
-                    operands_=[],
-                    asm_string="s_waitcnt lgkmcnt(0)",
-                    constraints="",
-                    has_side_effects=True,
-                )
-
-                # Main K-loop (identical to NN baseline).
-                for k in range_constexpr(K_ITERS - 2):
-                    b0_frag = b_s2r.load(b_cur0)
-                    a0_frag = a_s2r.load(a_cur0)
-                    a_g2s.load(a_next1, A1_gl_offset + (k + 1) * BLOCK_K)
-                    rocdl.s_barrier()
-
-                    rocdl.s_setprio(1)
-                    c00_frag = mfma.call(a0_frag, b0_frag, c00_frag)
-                    rocdl.s_setprio(0)
-                    rocdl.s_barrier()
-
-                    b1_frag = b_s2r.load(b_cur1)
-                    b_g2s.load(b_cur0, B0_gl_offset + (k + 2) * BLOCK_K * c_n)
-                    rocdl.s_barrier()
-
-                    rocdl.s_setprio(1)
-                    c01_frag = mfma.call(a0_frag, b1_frag, c01_frag)
-                    rocdl.s_setprio(0)
-                    rocdl.s_barrier()
-
-                    a1_frag = a_s2r.load(a_cur1)
-                    a_g2s.load(a_cur0, A0_gl_offset + (k + 2) * BLOCK_K)
-                    rocdl.s_barrier()
-
-                    rocdl.s_setprio(1)
-                    c10_frag = mfma.call(a1_frag, b0_frag, c10_frag)
-                    rocdl.s_setprio(0)
-                    rocdl.s_barrier()
-
-                    b_g2s.load(b_cur1, B1_gl_offset + (k + 2) * BLOCK_K * c_n)
-                    wait_barrier(2 * N_LDS_STEPS_A + N_LDS_STEPS_B)
-
-                    rocdl.s_setprio(1)
-                    c11_frag = mfma.call(a1_frag, b1_frag, c11_frag)
-                    rocdl.s_setprio(0)
-                    rocdl.s_barrier()
-
-                    a_cur0, a_next0 = a_next0, a_cur0
-                    a_cur1, a_next1 = a_next1, a_cur1
-                    b_cur0, b_next0 = b_next0, b_cur0
-                    b_cur1, b_next1 = b_next1, b_cur1
-
-                # Epilog 1.
-                k = K_ITERS - 2
-                b0_frag = b_s2r.load(b_cur0)
-                a0_frag = a_s2r.load(a_cur0)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c00_frag = mfma.call(a0_frag, b0_frag, c00_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                b1_frag = b_s2r.load(b_cur1)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c01_frag = mfma.call(a0_frag, b1_frag, c01_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                a1_frag = a_s2r.load(a_cur1)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c10_frag = mfma.call(a1_frag, b0_frag, c10_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                b0_frag = b_s2r.load(b_next0)
-                a_g2s.load(a_next1, A1_gl_offset + (k + 1) * BLOCK_K)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c11_frag = mfma.call(a1_frag, b1_frag, c11_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                a_cur0, a_next0 = a_next0, a_cur0
-                a_cur1, a_next1 = a_next1, a_cur1
-                b_cur0, b_next0 = b_next0, b_cur0
-                b_cur1, b_next1 = b_next1, b_cur1
-
-                # Epilog 2.
-                a0_frag = a_s2r.load(a_cur0)
-                wait_barrier(0)
-
-                rocdl.s_setprio(1)
-                c00_frag = mfma.call(a0_frag, b0_frag, c00_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                b1_frag = b_s2r.load(b_cur1)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c01_frag = mfma.call(a0_frag, b1_frag, c01_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                a1_frag = a_s2r.load(a_cur1)
-                rocdl.s_barrier()
-
-                rocdl.s_setprio(1)
-                c10_frag = mfma.call(a1_frag, b0_frag, c10_frag)
-                c11_frag = mfma.call(a1_frag, b1_frag, c11_frag)
-                rocdl.s_setprio(0)
-                rocdl.s_barrier()
-
-                wave_n_offset = wave_n * (N_TILES_B * 16)
-                wave_m_offset = wave_m * (N_TILES_A * 16)
-                base_row = block_m * BLOCK_M + wave_m_offset
-                base_col = block_n * BLOCK_N + wave_n_offset
-
-                store_c.store(c00_frag, base_row + 0, base_col + 0)
-                store_c.store(c01_frag, base_row + 0, base_col + LDS_BLOCK_N)
-                store_c.store(c10_frag, base_row + LDS_BLOCK_M, base_col + 0)
-                store_c.store(c11_frag, base_row + LDS_BLOCK_M, base_col + LDS_BLOCK_N)
-
-        @flyc.jit
-        def launch_dense_nn_persistent(
-            A: fx.Tensor,
-            B: fx.Tensor,
-            C: fx.Tensor,
-            A_scale: fx.Tensor,
-            B_scale: fx.Tensor,
-            c_m: fx.Int32,
-            c_n: fx.Int32,
-            stream: fx.Stream,
-        ):
-            num_tiles = ceildiv(c_m, BLOCK_M) * ceildiv(c_n, BLOCK_N)
-            grid_x = ceildiv(num_tiles, PERSIST_FACTOR)
-            kernel_dense_nn_persistent(
-                A,
-                B,
-                C,
-                A_scale,
-                B_scale,
-                c_m,
-                c_n,
-                value_attrs={"rocdl.waves_per_eu": 2, "rocdl.flat_work_group_size": "512,512"},
-            ).launch(grid=(grid_x, 1, 1), block=(512, 1, 1), stream=stream)
-
-        return launch_dense_nn_persistent
-
-    # ──────────────────────────────────────────────────────────────────────
-    # Session C3: CRR (TN/wgrad) dense kernel.
-    # C[k_out, n_out] = sum_{m_contract} A[m_contract, k_out] * B[m_contract, n_out]
-    # A storage [M_contract, K_out] row-major K_out-contig.
-    # B storage [M_contract, N_out] row-major N_out-contig (== NN B byte-identical).
-    # C storage [K_out, N_out] row-major N_out-contig.
-    # mfma view: mfma_M = K_out (16), mfma_K = M_contract (128), mfma_N = N_out (16).
-    # Both A and B need transpose-load (contraction = storage outer dim) → both
-    # use S2RLoaderTr; A-path g2s uses compute_global_swizzle_nn with K_out
-    # as the row-stride (mirror of NN B with K_inner → M_contract rename).
-    # Probe verified: scripts2/probe_transpose_crr_{a,b}_lds.py (3-leg PASS).
-    # Kernel structure is a 1:1 mirror of `_compile_dense_nn` with:
-    #   - A_gl_offset uses block_m * BLOCK_M (column-start, like NN B), not
-    #     (block_m * BLOCK_M) * K
-    #   - A K-iter additive = k * BLOCK_K * c_m (row-stride × k), not k * BLOCK_K
-    #   - A path uses compute_global_swizzle_nn (stride c_m) + S2RLoaderTr(wave_m)
-    # ──────────────────────────────────────────────────────────────────────
-    @functools.lru_cache(maxsize=128)
-    def _compile_dense_crr(
-        K: int,
-        BLOCK_M: int = 256,
-        BLOCK_N: int = 256,
-        WAVES_PER_EU: int = 2,
-        SETPRIO_HIGH: int = 1,
-        SCHED_HINT: int = 0,
-        GROUP_M: int = 1,
-        MAXNREG: int = 0,
-    ):
-        """CRR-layout fp8 dense kernel (TN/wgrad).
-
-        Param naming kept symmetric with NN factory:
-          K            - M_contract (contraction extent); compile-time, K % 128 == 0
-          BLOCK_M      - K_out tile (mfma_M direction), default 256
-          BLOCK_N      - N_out tile (mfma_N direction), default 256
-          BLOCK_K      - M_contract tile, fixed 128 (mfma K size)
-
-        Kernel signature inside FlyDSL kernel:
-          c_m = K_out   (output row dim, C is [c_m, c_n])
-          c_n = N_out   (output col dim)
-        """
-        BLOCK_K = 128
-        assert BLOCK_M >= 128 and BLOCK_N >= 256 and BLOCK_M % 128 == 0 and BLOCK_N % 256 == 0
-        assert K % BLOCK_K == 0
-        assert GROUP_M >= 1
-
-        K_ITERS = K // BLOCK_K
-        assert K_ITERS >= 2
-
-        N_TILES_A = BLOCK_M // 64
-        N_TILES_B = BLOCK_N // 128
-        N_ACCUMS = N_TILES_A * N_TILES_B
-        LDS_BLOCK_M = BLOCK_M // 2
-        LDS_BLOCK_N = BLOCK_N // 2
-        N_LDS_STEPS_A = LDS_BLOCK_M // 64
-        N_LDS_STEPS_B = LDS_BLOCK_N // 64
-        N_LDS_ROUNDS = max(N_LDS_STEPS_A, N_LDS_STEPS_B)
-        a_lds_size = LDS_BLOCK_M * BLOCK_K
-        b_lds_size = LDS_BLOCK_N * BLOCK_K
-
-        @fx.struct
-        class SharedStorage:
-            A_lds_cur_0: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            A_lds_cur_1: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            A_lds_next_0: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            A_lds_next_1: fx.Array[fx.Float8E4M3FN, a_lds_size, 16]
-            B_lds_cur_0: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-            B_lds_cur_1: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-            B_lds_next_0: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-            B_lds_next_1: fx.Array[fx.Float8E4M3FN, b_lds_size, 16]
-
-        @flyc.kernel(known_block_size=[512, 1, 1])
-        def kernel_dense_crr(
-            A: fx.Tensor,
-            B: fx.Tensor,
-            C: fx.Tensor,
-            A_scale: fx.Tensor,
-            B_scale: fx.Tensor,
-            c_m: fx.Int32,
-            c_n: fx.Int32,
-        ):
-            # Trace-order workaround (mirror NN; see _compile_dense_nn comment).
-            _ = str(fx.thread_idx.x)
-            F8_IR_t = fx.Float8E4M3FN.ir_type
-
-            n_blocks = ceildiv(c_n, BLOCK_N)
-
-            lds = fx.SharedAllocator().allocate(SharedStorage).peek()
-            a_cur0 = lds.A_lds_cur_0
-            a_cur1 = lds.A_lds_cur_1
-            a_next0 = lds.A_lds_next_0
-            a_next1 = lds.A_lds_next_1
-            b_cur0 = lds.B_lds_cur_0
-            b_cur1 = lds.B_lds_cur_1
-            b_next0 = lds.B_lds_next_0
-            b_next1 = lds.B_lds_next_1
-
-            lane_id = fx.thread_idx.x % 64
-            wave_id = fx.thread_idx.x // 64
-            wave_m = wave_id // 4
-            wave_n = wave_id % 4
-            pid = fx.block_idx.x
-            num_pid_in_group = GROUP_M * n_blocks
-            group_id = pid // num_pid_in_group
-            pid_in_group = pid % num_pid_in_group
-            pid_m_inner = pid_in_group % GROUP_M
-            pid_n = pid_in_group // GROUP_M
-            block_m = group_id * GROUP_M + pid_m_inner
-            block_n = pid_n
-
-            # CRR A: [M_contract, K_out] row-major K_out-contig. We hold a
-            # fixed K_out tile [block_m*BLOCK_M : ... + BLOCK_M] and stream
-            # M_contract rows. Per-WG starts are K_out column positions.
-            A0_gl_offset = block_m * BLOCK_M + 0
-            A1_gl_offset = block_m * BLOCK_M + LDS_BLOCK_M
-
-            # CRR B: [M_contract, N_out] row-major N_out-contig. Identical
-            # in form to NN B (M_contract == K_inner). Per-WG starts are
-            # N_out column positions.
-            B0_gl_offset = block_n * BLOCK_N + 0
-            B1_gl_offset = block_n * BLOCK_N + LDS_BLOCK_N
-
-            gA = make_fp8_buffer_tensor(A, F8_IR_t)
-            gB = make_fp8_buffer_tensor(B, F8_IR_t)
-            a_div = fx.logical_divide(gA, fx.make_layout(1, 1))
-            b_div = fx.logical_divide(gB, fx.make_layout(1, 1))
-
-            # A path g2s: same shape as NN B's g2s — stride = K_out = c_m.
-            gl_off_a = compute_global_swizzle_nn(lane_id, wave_id, c_m, N_LDS_ROUNDS)
-            # B path g2s: NN B's exact helper, stride = N_out = c_n.
-            gl_off_b = compute_global_swizzle_nn(lane_id, wave_id, c_n, N_LDS_ROUNDS)
-
-            mfma = Mfma16x16x128(N_TILES_A, N_TILES_B)
-
-            a_g2s = G2SLoader(a_div, gl_off_a, N_LDS_STEPS_A, F8_IR_t, wave_id)
-            b_g2s = G2SLoader(b_div, gl_off_b, N_LDS_STEPS_B, F8_IR_t, wave_id)
-            # A path uses S2RLoaderTrA (parameterized wave-stride =
-            # n_tiles_a * 16; needed because wave_m ∈ [0,2) with N_TILES_A=4
-            # requires stride 64, not S2RLoaderTr's hardcoded 32 which was
-            # specialized for NN B's wave_n ∈ [0,4) × N_TILES_B=2 case).
-            # B path = NN B direct reuse (S2RLoaderTr).  Per CRR C1/C2
-            # probes (32768/32768 PASS).
-            a_s2r = S2RLoaderTrA(wave_m, N_TILES_A)
-            b_s2r = S2RLoaderTr(wave_n, N_TILES_B)
-            store_c = StoreC(A_scale, B_scale, C, c_m, c_n, mfma.idx, N_TILES_A, N_TILES_B)
-
-            c00_frag = [mfma.zero_value] * N_ACCUMS
-            c01_frag = [mfma.zero_value] * N_ACCUMS
-            c10_frag = [mfma.zero_value] * N_ACCUMS
-            c11_frag = [mfma.zero_value] * N_ACCUMS
-
-            _CRR_SP = SETPRIO_HIGH
-            _CRR_SH = SCHED_HINT
-
-            def _prio_raise():
-                if _CRR_SP > 0:
-                    rocdl.s_setprio(_CRR_SP)
-
-            def _prio_drop():
-                if _CRR_SP > 0:
-                    rocdl.s_setprio(0)
-
-            def _sched_after_mfma(idx=None):
-                if _CRR_SH == 1:
-                    rocdl.sched_group_barrier(rocdl.mask_mfma, N_ACCUMS, 0)
-                elif _CRR_SH == 2:
-                    rocdl.sched_group_barrier(rocdl.mask_mfma, N_ACCUMS, 0)
-                    rocdl.sched_group_barrier(rocdl.mask_vmem_rd, 1, 1)
-                    rocdl.sched_group_barrier(rocdl.mask_dsrd, 1, 2)
-                elif _CRR_SH == 3 and idx == 1:
-                    rocdl.sched_group_barrier(rocdl.mask_mfma, N_ACCUMS, 0)
-                elif _CRR_SH == 4 and idx == 0:
-                    rocdl.sched_group_barrier(rocdl.mask_mfma, N_ACCUMS, 0)
-                elif _CRR_SH == 5 and idx == 2:
-                    rocdl.sched_group_barrier(rocdl.mask_mfma, N_ACCUMS, 0)
-
-            # Prelude.  A K-iter step in flat A units = BLOCK_K * c_m
-            # (M_contract advances by BLOCK_K, with K_out=c_m as row stride).
-            # B K-iter step in flat B units = BLOCK_K * c_n (same as NN).
-            b_g2s.load(b_cur0, B0_gl_offset + 0 * BLOCK_K * c_n)
-            a_g2s.load(a_cur0, A0_gl_offset + 0 * BLOCK_K * c_m)
-            b_g2s.load(b_cur1, B1_gl_offset + 0 * BLOCK_K * c_n)
-            a_g2s.load(a_cur1, A1_gl_offset + 0 * BLOCK_K * c_m)
-
-            if wave_m == 1:
-                rocdl.s_barrier()
-
-            wait_barrier(N_LDS_STEPS_A + N_LDS_STEPS_B)
-
-            b_g2s.load(b_next0, B0_gl_offset + 1 * BLOCK_K * c_n)
-            a_g2s.load(a_next0, A0_gl_offset + 1 * BLOCK_K * c_m)
-            b_g2s.load(b_next1, B1_gl_offset + 1 * BLOCK_K * c_n)
-
-            wait_barrier(N_LDS_STEPS_A + 2 * N_LDS_STEPS_B)
-
-            # Main loop (mirror NN ping-pong).
-            for k in range_constexpr(K_ITERS - 2):
-                b0_frag = b_s2r.load(b_cur0)
-                a0_frag = a_s2r.load(a_cur0)
-                a_g2s.load(a_next1, A1_gl_offset + (k + 1) * BLOCK_K * c_m)
-                rocdl.s_barrier()
-
-                _prio_raise()
-                c00_frag = mfma.call(a0_frag, b0_frag, c00_frag)
-                _prio_drop()
-                _sched_after_mfma(0)
-                rocdl.s_barrier()
-
-                b1_frag = b_s2r.load(b_cur1)
-                b_g2s.load(b_cur0, B0_gl_offset + (k + 2) * BLOCK_K * c_n)
-                rocdl.s_barrier()
-
-                _prio_raise()
-                c01_frag = mfma.call(a0_frag, b1_frag, c01_frag)
-                _prio_drop()
-                _sched_after_mfma(1)
-                rocdl.s_barrier()
-
-                a1_frag = a_s2r.load(a_cur1)
-                a_g2s.load(a_cur0, A0_gl_offset + (k + 2) * BLOCK_K * c_m)
-                rocdl.s_barrier()
-
-                _prio_raise()
-                c10_frag = mfma.call(a1_frag, b0_frag, c10_frag)
-                _prio_drop()
-                _sched_after_mfma(2)
-                rocdl.s_barrier()
-
-                b_g2s.load(b_cur1, B1_gl_offset + (k + 2) * BLOCK_K * c_n)
-                wait_barrier(2 * N_LDS_STEPS_A + N_LDS_STEPS_B)
-
-                _prio_raise()
-                c11_frag = mfma.call(a1_frag, b1_frag, c11_frag)
-                _prio_drop()
-                _sched_after_mfma(3)
-                rocdl.s_barrier()
-
-                a_cur0, a_next0 = a_next0, a_cur0
-                a_cur1, a_next1 = a_next1, a_cur1
-                b_cur0, b_next0 = b_next0, b_cur0
-                b_cur1, b_next1 = b_next1, b_cur1
-
-            # Epilog 1.
-            k = K_ITERS - 2
-            b0_frag = b_s2r.load(b_cur0)
-            a0_frag = a_s2r.load(a_cur0)
-            rocdl.s_barrier()
-
-            rocdl.s_setprio(1)
-            c00_frag = mfma.call(a0_frag, b0_frag, c00_frag)
-            rocdl.s_setprio(0)
-            rocdl.s_barrier()
-
-            b1_frag = b_s2r.load(b_cur1)
-            rocdl.s_barrier()
-
-            rocdl.s_setprio(1)
-            c01_frag = mfma.call(a0_frag, b1_frag, c01_frag)
-            rocdl.s_setprio(0)
-            rocdl.s_barrier()
-
-            a1_frag = a_s2r.load(a_cur1)
-            rocdl.s_barrier()
-
-            rocdl.s_setprio(1)
-            c10_frag = mfma.call(a1_frag, b0_frag, c10_frag)
-            rocdl.s_setprio(0)
-            rocdl.s_barrier()
-
-            b0_frag = b_s2r.load(b_next0)
-            a_g2s.load(a_next1, A1_gl_offset + (k + 1) * BLOCK_K * c_m)  # stale-a1 fix
-            rocdl.s_barrier()
-
-            rocdl.s_setprio(1)
-            c11_frag = mfma.call(a1_frag, b1_frag, c11_frag)
-            rocdl.s_setprio(0)
-            rocdl.s_barrier()
-
-            a_cur0, a_next0 = a_next0, a_cur0
-            a_cur1, a_next1 = a_next1, a_cur1
-            b_cur0, b_next0 = b_next0, b_cur0
-            b_cur1, b_next1 = b_next1, b_cur1
-
-            # Epilog 2.
-            a0_frag = a_s2r.load(a_cur0)
-            wait_barrier(0)
-
-            rocdl.s_setprio(1)
-            c00_frag = mfma.call(a0_frag, b0_frag, c00_frag)
-            rocdl.s_setprio(0)
-            rocdl.s_barrier()
-
-            b1_frag = b_s2r.load(b_cur1)
-            rocdl.s_barrier()
-
-            rocdl.s_setprio(1)
-            c01_frag = mfma.call(a0_frag, b1_frag, c01_frag)
-            rocdl.s_setprio(0)
-            rocdl.s_barrier()
-
-            a1_frag = a_s2r.load(a_cur1)
-            rocdl.s_barrier()
-
-            rocdl.s_setprio(1)
-            c10_frag = mfma.call(a1_frag, b0_frag, c10_frag)
-            c11_frag = mfma.call(a1_frag, b1_frag, c11_frag)
-            rocdl.s_setprio(0)
-            rocdl.s_barrier()
-
-            wave_n_offset = wave_n * (N_TILES_B * 16)
-            wave_m_offset = wave_m * (N_TILES_A * 16)
-            base_row = block_m * BLOCK_M + wave_m_offset
-            base_col = block_n * BLOCK_N + wave_n_offset
-
-            store_c.store(c00_frag, base_row + 0, base_col + 0)
-            store_c.store(c01_frag, base_row + 0, base_col + LDS_BLOCK_N)
-            store_c.store(c10_frag, base_row + LDS_BLOCK_M, base_col + 0)
-            store_c.store(c11_frag, base_row + LDS_BLOCK_M, base_col + LDS_BLOCK_N)
-
-        @flyc.jit
-        def launch_dense_crr(
-            A: fx.Tensor,
-            B: fx.Tensor,
-            C: fx.Tensor,
-            A_scale: fx.Tensor,
-            B_scale: fx.Tensor,
-            c_m: fx.Int32,
-            c_n: fx.Int32,
-            stream: fx.Stream,
-        ):
-            grid_x = ceildiv(c_m, BLOCK_M) * ceildiv(c_n, BLOCK_N)
-            kernel_dense_crr(
-                A,
-                B,
-                C,
-                A_scale,
-                B_scale,
-                c_m,
-                c_n,
-                value_attrs=(
-                    {
-                        "rocdl.waves_per_eu": WAVES_PER_EU,
-                        "rocdl.flat_work_group_size": "512,512",
-                        "passthrough": [["amdgpu-num-vgpr", str(MAXNREG)]],
-                    }
-                    if MAXNREG > 0
-                    else {"rocdl.waves_per_eu": WAVES_PER_EU, "rocdl.flat_work_group_size": "512,512"}
-                ),
-            ).launch(grid=(grid_x, 1, 1), block=(512, 1, 1), stream=stream)
-
-        return launch_dense_crr
-
 else:
 
     def _compile_dense_nt(K: int, BLOCK_M: int = 256, BLOCK_N: int = 256):
-        raise ImportError("FlyDSL is not available -- this entry should be gated by can_handle.")
-
-    def _compile_dense_nt_persistent(
-        K: int, BLOCK_M: int = 256, BLOCK_N: int = 256, PERSIST_FACTOR: int = 2,
-        A11_SAMETILE: bool = False,
-    ):
         raise ImportError("FlyDSL is not available -- this entry should be gated by can_handle.")
 
     def _compile_dense_nn(
@@ -1968,24 +861,6 @@ else:
         SETPRIO_HIGH: int = 1,
         SCHED_HINT: int = 0,
         GROUP_M: int = 1,
-    ):
-        raise ImportError("FlyDSL is not available -- this entry should be gated by can_handle.")
-
-    def _compile_dense_nn_persistent(
-        K: int, BLOCK_M: int = 256, BLOCK_N: int = 256, PERSIST_FACTOR: int = 2,
-        B7_SAMETILE: bool = False,
-    ):
-        raise ImportError("FlyDSL is not available -- this entry should be gated by can_handle.")
-
-    def _compile_dense_crr(
-        K: int,
-        BLOCK_M: int = 256,
-        BLOCK_N: int = 256,
-        WAVES_PER_EU: int = 2,
-        SETPRIO_HIGH: int = 1,
-        SCHED_HINT: int = 0,
-        GROUP_M: int = 1,
-        MAXNREG: int = 0,
     ):
         raise ImportError("FlyDSL is not available -- this entry should be gated by can_handle.")
 
@@ -2124,12 +999,6 @@ def _pick_group_m_nn(N: int, K: int) -> int:
     return _NN_GROUP_M_OVERRIDE.get((N, K), _NN_GROUP_M_DEFAULT)
 
 
-# Session A7 persistent-NT knob. Default 0 = non-persistent baseline (no
-# behavior change). Set FLYDSL_RCR_PERSIST_FACTOR=2 to opt-in. Kernel-only KPI
-# gate: worst shape qwen_down_B16_M4096 fly/hbl ≥ 1.00. Only activated when:
-#   - GROUP_M == 1 (don't mix with super-block swizzle)
-#   - num_tiles % PF == 0 (host-side gate; falls back to non-persistent otherwise)
-_NT_PERSIST_FACTOR = int(os.environ.get("FLYDSL_RCR_PERSIST_FACTOR", "0"))
 # Phase B2 NN-path sched/setprio/waves_per_eu sweep knobs. Default values
 # reproduce the B1 baseline kernel (waves_per_eu=2, setprio raise/drop pair,
 # no sched_group_barrier hint). Variants are evaluated by `_b2_run_sweep.sh`
@@ -2155,81 +1024,8 @@ _NN_SCHED_HINT = int(os.environ.get("FLYDSL_NN_SCHED_HINT", "0"))
 # plumbing kept for future use (or to study scratch-spill perf curve);
 # default 0 has 0 overhead and reproduces B3 baseline bit-exactly.
 _NN_MAXNREG = int(os.environ.get("FLYDSL_NN_MAXNREG", "0"))
-# A11 sametile probe: when 1 AND PF>=2, iter1+ uses iter0's pid (drops
-# tile_iv from coord). Only meaningful for diagnostic; correctness is
-# "iter1 overwrites iter0 with same coords" so result is undefined for
-# non-multiple-of-PF num_tiles. Use scripts2/_a11_same_tile_diag.py.
-_NT_A11_SAMETILE = int(os.environ.get("FLYDSL_RCR_A11_SAMETILE", "0")) == 1
-
-# Session B7 persistent-NN knobs (mirror NT). Default 0 = non-persistent
-# baseline (no behavior change). Set FLYDSL_NN_PERSIST_FACTOR=2 to opt-in.
-# B7_SAMETILE probe: when 1 AND PF>=2, iter1+ drops tile_iv from pid.
-# Only activated when:
-#   - GROUP_M == 1 (don't mix with super-block swizzle; B3 default GM=4
-#     so probe MUST set FLYDSL_NN_GROUP_M_FORCE=1)
-#   - num_tiles % PF == 0 (host-side gate; falls back to non-persistent otherwise)
-_NN_PERSIST_FACTOR = int(os.environ.get("FLYDSL_NN_PERSIST_FACTOR", "0"))
-_NN_B7_SAMETILE = int(os.environ.get("FLYDSL_NN_B7_SAMETILE", "0")) == 1
-
-# Session C3: CRR (TN/wgrad) knobs. Mirror NN defaults; C3 is a skeleton/G0
-# correctness gate (no perf sweeps yet — those land in C4/C5/C6).
-_CRR_WAVES_PER_EU = int(os.environ.get("FLYDSL_CRR_WAVES_PER_EU", "2"))
-_CRR_SETPRIO_HIGH = int(os.environ.get("FLYDSL_CRR_SETPRIO_HIGH", "1"))
-_CRR_SCHED_HINT = int(os.environ.get("FLYDSL_CRR_SCHED_HINT", "0"))
-_CRR_MAXNREG = int(os.environ.get("FLYDSL_CRR_MAXNREG", "0"))
-_CRR_GROUP_M_OVERRIDE: dict[tuple[int, int], int] = {
-    # C5.2 GROUP_M sweep wins on large-M_contract shapes (worst-tier per C4 baseline).
-    # Key = (N_out, M_contract). Picked by single-shape min-of-3-trial ratio vs Triton TN.
-    # Safety: _pick_group_m_crr guards against K_out tile count not divisible by GROUP_M
-    # (kernel dispatcher assumes evenly divisible — otherwise OOB writes corrupt C).
-    (4096, 32768): 4,  # dsv3 up M_c=32768  K_o=7168 N_o=4096 : +6.3pp vs GM=1
-    (4096, 65536): 4,  # dsv3 up M_c=65536  K_o=7168 N_o=4096 : +2.3pp vs GM=1
-    (3072, 65536): 4,  # qwen up M_c=65536  K_o=4096 N_o=3072 : +1.8pp vs GM=1
-    # C6.1 GROUP_M sweep wins (Δ vs GM=1 > 1.5pp accept threshold)
-    (7168,  8192): 2,  # dsv3 down M_c=8192   K_o=2048 N_o=7168 : +2.11pp (M_tiles=8 %2=0)
-    (4096, 16384): 2,  # dsv3 up   M_c=16384  K_o=7168 N_o=4096 : +3.01pp (M_tiles=28 %2=0)
-    (3072,  8192): 8,  # qwen up   M_c=8192   K_o=4096 N_o=3072 : +3.98pp (M_tiles=16 %8=0)
-    (3072, 16384): 4,  # qwen up   M_c=16384  K_o=4096 N_o=3072 : +3.78pp (M_tiles=16 %4=0)
-    (3072, 32768): 4,  # qwen up   M_c=32768  K_o=4096 N_o=3072 : +3.44pp (M_tiles=16 %4=0)
-}
-_CRR_GROUP_M_DEFAULT = 1
-_CRR_GROUP_M_FORCE = int(os.environ.get("FLYDSL_CRR_GROUP_M_FORCE", "0"))
-_CRR_BLOCK_M = 256  # kept in sync with _compile_dense_crr BLOCK_M default
 
 
-def _pick_group_m_crr(N: int, K: int, K_out: int = 0) -> int:
-    raw = _CRR_GROUP_M_FORCE if _CRR_GROUP_M_FORCE > 0 else \
-          _CRR_GROUP_M_OVERRIDE.get((N, K), _CRR_GROUP_M_DEFAULT)
-    if raw <= 1 or K_out <= 0:
-        return raw
-    # Safety: kernel super-block dispatcher requires M_tiles % GROUP_M == 0.
-    # Different K_out values can share the same (N_out, M_contract) key, so
-    # callers must pass K_out and we fall back to 1 if grouping would over-launch
-    # past the valid M-tile range (would write OOB → NaN/corruption).
-    m_tiles = (K_out + _CRR_BLOCK_M - 1) // _CRR_BLOCK_M
-    if m_tiles % raw != 0:
-        return 1
-    return raw
-
-
-def _pick_persist_factor(num_tiles: int, group_m: int) -> int:
-    if _NT_PERSIST_FACTOR <= 1:
-        return 0
-    if group_m != 1:
-        return 0
-    if num_tiles % _NT_PERSIST_FACTOR != 0:
-        return 0
-    return _NT_PERSIST_FACTOR
-
-
-def _pick_persist_factor_nn(num_tiles: int, group_m: int) -> int:
-    if _NN_PERSIST_FACTOR <= 1:
-        return 0
-    if group_m != 1:
-        return 0
-    if num_tiles % _NN_PERSIST_FACTOR != 0:
-        return 0
-    return _NN_PERSIST_FACTOR
 
 
 def gemm_fp8_tensorwise_flydsl_kernel(
@@ -2248,17 +1044,21 @@ def gemm_fp8_tensorwise_flydsl_kernel(
       - NT (trans_a=F, trans_b=T): native NT kernel, no host transpose.
       - NN (trans_a=F, trans_b=F): native NN kernel (used by dgrad path), no
         host transpose; uses ds_read_b64_tr_b8 for B-operand load.
-      - CRR/TN (trans_a=T, trans_b=F): native CRR kernel (used by wgrad path),
-        no host transpose; uses ds_read_b64_tr_b8 for BOTH A and B (both have
-        contraction as storage outer dim). Session C3.
       - TT: host-canonicalised to NT (transpose A and B), then NT kernel.
-    trans_c=True returned as post-hoc out.t().contiguous() (extra mem copy
-    vs Triton which writes [N, M] directly via swapped strides).
+      - TN: not supported (raises). For wgrad use grouped path.
+    trans_c=True returned as post-hoc out.t().contiguous().
     """
     if out_dtype != torch.bfloat16:
         raise NotImplementedError(f"FlyDSL wrapper only emits bf16. Got {out_dtype}.")
     assert a.dim() == 2 and b.dim() == 2
 
+    if trans_a and (not trans_b):
+        raise NotImplementedError(
+            "FlyDSL dense fp8 GEMM no longer supports TN/CRR layout in this module. "
+            "Use the grouped-gemm path for wgrad."
+        )
+
+    # Dispatch by layout (NN native path skips canonicalisation).
     # Dispatch by layout (NN native path skips canonicalisation).
     if (not trans_a) and (not trans_b):
         # NN native: A [M, K], B [K, N].
@@ -2274,22 +1074,14 @@ def gemm_fp8_tensorwise_flydsl_kernel(
         b_scale_v = _broadcast_scale(b_scale_inv, N, a.device)
         out = torch.empty((M, N), dtype=out_dtype, device=a.device)
         gm_nn = _pick_group_m_nn(N, K)
-        num_tiles_nn = ((M + 255) // 256) * ((N + 255) // 256)
-        pf_nn = _pick_persist_factor_nn(num_tiles_nn, gm_nn)
-        if pf_nn > 0:
-            launch = _compile_dense_nn_persistent(
-                K=K, BLOCK_M=256, BLOCK_N=256, PERSIST_FACTOR=pf_nn,
-                B7_SAMETILE=_NN_B7_SAMETILE,
-            )
-        else:
-            launch = _compile_dense_nn(
-                K=K, BLOCK_M=256, BLOCK_N=256,
-                WAVES_PER_EU=_NN_WAVES_PER_EU,
-                SETPRIO_HIGH=_NN_SETPRIO_HIGH,
-                SCHED_HINT=_NN_SCHED_HINT,
-                GROUP_M=gm_nn,
-                MAXNREG=_NN_MAXNREG,
-            )
+        launch = _compile_dense_nn(
+            K=K, BLOCK_M=256, BLOCK_N=256,
+            WAVES_PER_EU=_NN_WAVES_PER_EU,
+            SETPRIO_HIGH=_NN_SETPRIO_HIGH,
+            SCHED_HINT=_NN_SCHED_HINT,
+            GROUP_M=gm_nn,
+            MAXNREG=_NN_MAXNREG,
+        )
         args = (
             _as_i8_flat(a),
             _as_i8_flat(b),
@@ -2301,44 +1093,6 @@ def gemm_fp8_tensorwise_flydsl_kernel(
             torch.cuda.current_stream(),
         )
         _get_compiled_dense(launch, args, maxnreg=_NN_MAXNREG)(*args)
-    elif trans_a and (not trans_b):
-        # CRR native (TN / wgrad): A [M_contract, K_out], B [M_contract, N_out],
-        # C = A^T @ B has shape [K_out, N_out]. Both A and B need transpose-load
-        # because contraction = storage outer dim. Probe verified at
-        # scripts2/probe_transpose_crr_{a,b}_lds.py (32768/32768 bytes PASS).
-        M_a, K_out = a.shape
-        M_b, N_out = b.shape
-        assert M_a == M_b, f"CRR contraction-M mismatch: a {a.shape}, b {b.shape}"
-        M_contract = M_a
-        if M_contract % 128 != 0:
-            raise NotImplementedError(
-                f"FlyDSL CRR requires M_contract % 128 == 0 (BLOCK_K=128). "
-                f"Got M_contract={M_contract}."
-            )
-        # Tensorwise: scale is scalar (numel==1), broadcast to output dims.
-        a_scale_v = _broadcast_scale(a_scale_inv, K_out, a.device)
-        b_scale_v = _broadcast_scale(b_scale_inv, N_out, a.device)
-        out = torch.empty((K_out, N_out), dtype=out_dtype, device=a.device)
-        gm_crr = _pick_group_m_crr(N_out, M_contract, K_out)
-        launch = _compile_dense_crr(
-            K=M_contract, BLOCK_M=256, BLOCK_N=256,
-            WAVES_PER_EU=_CRR_WAVES_PER_EU,
-            SETPRIO_HIGH=_CRR_SETPRIO_HIGH,
-            SCHED_HINT=_CRR_SCHED_HINT,
-            GROUP_M=gm_crr,
-            MAXNREG=_CRR_MAXNREG,
-        )
-        args = (
-            _as_i8_flat(a),
-            _as_i8_flat(b),
-            out.contiguous().view(-1),
-            a_scale_v,
-            b_scale_v,
-            K_out,
-            N_out,
-            torch.cuda.current_stream(),
-        )
-        _get_compiled_dense(launch, args, maxnreg=_CRR_MAXNREG)(*args)
     else:
         # NT native OR TT via host canonicalisation.
         a_nt, b_nt, M, N, K = _canonicalize_nt(a, b, trans_a, trans_b)
@@ -2350,15 +1104,7 @@ def gemm_fp8_tensorwise_flydsl_kernel(
         b_scale_v = _broadcast_scale(b_scale_inv, N, a.device)
         out = torch.empty((M, N), dtype=out_dtype, device=a.device)
         gm = _pick_group_m(N, K)
-        num_tiles = ((M + 255) // 256) * ((N + 255) // 256)
-        pf = _pick_persist_factor(num_tiles, gm)
-        if pf > 0:
-            launch = _compile_dense_nt_persistent(
-                K=K, BLOCK_M=256, BLOCK_N=256, PERSIST_FACTOR=pf,
-                A11_SAMETILE=_NT_A11_SAMETILE,
-            )
-        else:
-            launch = _compile_dense_nt(K=K, BLOCK_M=256, BLOCK_N=256, GROUP_M=gm)
+        launch = _compile_dense_nt(K=K, BLOCK_M=256, BLOCK_N=256, GROUP_M=gm)
         args = (
             _as_i8_flat(a_nt),
             _as_i8_flat(b_nt),
@@ -2370,7 +1116,6 @@ def gemm_fp8_tensorwise_flydsl_kernel(
             torch.cuda.current_stream(),
         )
         _get_compiled_dense(launch, args)(*args)
-
     if trans_c:
         return out.t().contiguous()
     return out
